Move loading picture logic to base activity
